requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
python-dotenv
langchain>=0.2.0
langchain-community>=0.0.10
langchain-openai>=0.0.6
chromadb>=0.4.22
